feat(entity): converted long to int

+++ OnlineGameStore.DAL/Entities/Game.cs
@@ -4,12 +4,12 @@ namespace OnlineGameStore.DAL;
 
 public class Game
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     [MaxLength(256)]
     public required string Name { get; set; } = string.Empty;
     [MaxLength(4096)]
     public required string Description { get; set; } = string.Empty;
-    public required long Publisher { get; set; }
-    public required long Genre { get; set; }
-    public required long License { get; set; }
+    public required int Publisher { get; set; }
+    public required int Genre { get; set; }
+    public required int License { get; set; }
 }
